LF-3582 rename ExpenseItemsForTypes to AddExpense
